LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/link-rel-stylesheet-referrerpolicy-no-referrer.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: non-speculative case did not fetch got disallowed value ""

    <!-- speculative case in document.write -->
    <link rel=stylesheet href="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=ff7c108f-699c-48ad-be65-19add4ef3007&amp;encodingcheck=&Gbreve;" referrerpolicy=no-referrer>



FAIL Speculative parsing, document.write(): link-rel-stylesheet-referrerpolicy-no-referrer Unhandled rejection: assert_not_equals: non-speculative case did not fetch got disallowed value ""
